multiple tag select & some validation issue solved

+++ app/Http/Controllers/TagsController.php
@@ -98,6 +98,12 @@ class TagsController extends Controller
      */
     public function destroy(Tag $tag)
     {
+        
+        if($tag->posts->count() > 0){
+            session()->flash('error', 'Tag cannot be deleted, because it is attach with some posts');
+            return redirect()->back();
+        }
+
         $tag->delete();
         session()->flash('success', 'Tag Deleted successfully');
         return redirect(route('tags.index'));
